Fix "pointer cast size mismatch" warnings on 32-bit arch's
Should hopefully resolve #444. Cast as uintptr_t from pointers and not
directly into a 64-bit type.

+++ src/core/nativecall_libffi.c
@@ -760,7 +760,7 @@ MVMObject * MVM_nativecall_invoke(MVMThreadContext *tc, MVMObject *res_type,
                     break;
                 case MVM_NATIVECALL_ARG_CPOINTER:
                     REPR(value)->box_funcs.set_int(tc, STABLE(value), value, OBJECT_BODY(value),
-                        (MVMint64)*(void **)*(void **)values[i]);
+                        (MVMint64)(uintptr_t)*(void **)*(void **)values[i]);
                     break;
                 default:
                     MVM_telemetry_interval_stop(tc, interval_id, "nativecall invoke failed");
